Pull npm-tests out so they're reusable
